Zombie targeting revamp
This commit fixes a number of issues...
1.) It fixes the really long targeting arrays caused by players
travelling in vehicles failing the "_unit in _targets" (because we add
the driver to targets, but we don't check for the driver, we check for
the player...)
2.) You can no longer distract zombies with broken bottles/etc if they
are already chasing a player. Also, Zombies will no longer get stuck
attacking dead players if there's someone alive nearby.
3.) Immense reductions in setVariable usage (setVariable are prioritized
just like publicVariables, thus we don't want to transmit them over the
network. unless we need to.)
2 hours/50 players with old system: ~46000 "targets" variable set (+
really long arrays, which makes it even worse)
2 hours/50 players with new system: ~12000 (No long arrays)

I'm porting this from DayZero, and i tried to make sure i diden't miss
anything but feel free to review the code so I don't introduce any bugs.